Create Readme.txt
CodeX is a German beverage company that is aiming to make its mark in the Indian market. 
A few months ago, they launched their energy drink in 10 cities in India. Their Marketing team is responsible for increasing brand awareness, market share, and product development. They conducted a survey in those 10 cities and received results from 10k respondents. As a data analyst, the task is to convert these survey results to meaningful insights which the team can use to drive actions.